fix: rename control_mode_manager.cpp to follow the correct convention
